bug: 66 - changed playheadColPrecise to float
git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@519 1c6c7da2-4a7c-0410-9345-fc53533f7dce

+++ juce/syzygryd_sequencer/Source/Models/OscOutput.cpp
@@ -134,7 +134,7 @@ void OscOutput::sendSync()
 			// XXX should there be the following in osc/OscOutboundPacketStream.cpp ?
 			//        OutboundPacketStream& OutboundPacketStream::operator<<( unsigned int rhs )
 			p << osc::BeginMessage ("/sync")
-			<< playheadColPrecise << ppqPosition << timeInSeconds << bpm
+			<< (float)playheadColPrecise << ppqPosition << timeInSeconds << bpm
 			<< panelIndex << tabIndex << numTabs << numRows << numCols
 			<< *blob
 			<< osc::EndMessage;
